added casting in read_params

+++ srcs/op_ldi.c
@@ -6,7 +6,7 @@
 /*   By: yabdulha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/10 15:59:15 by yabdulha          #+#    #+#             */
-/*   Updated: 2018/08/15 05:02:37 by yabdulha         ###   ########.fr       */
+/*   Updated: 2018/08/15 22:26:08 by yabdulha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	op_ldi(t_vm *vm, t_champ *champ)
 {
-	int	reg;
-	int	tmp;
+	intmax_t	reg;
+	int			tmp;
 
 	resolve_params(vm, champ, champ->params, 1 + 2);
 	resolve_indirect(vm, champ, 1);
@@ -27,6 +27,6 @@ void	op_ldi(t_vm *vm, t_champ *champ)
 		champ->carry = 1;
 	else
 		champ->carry = 0;
-	printf("p1: %d, p2: %d, p3: %d\n", champ->params->p1, champ->params->p2, champ->params->p3);
+ 	printf("p1: %d, p2: %d, p3: %d\n", champ->params->p1, champ->params->p2, champ->params->p3);
 	printf("\nPUT %d in REG %d\n", reg, champ->params->p3);
 }
